feat(cb2-2097): move config folder into src to fix invocation failures (#5)
* feat(cb2-2097): move config folder into src to resolve invocation failures
